+++ CHANGELOG.md
@@ -5,6 +5,7 @@ All notable changes to `laravel-mailcoach` will be documented in this file
 ## 2.11.6 - 2020-05-14
 
 - fix for sql_mode=only_full_group_by issue
+- fix campaign not being marked as sent with a custom segment
 
 ## 2.11.5 - 2020-05-07